requirements.txt
pydub
python-dotenv
tqdm
requests
faiss-cpu==1.11.0
sentence-transformers
openai-whisper
pyaudio
TTS
sounddevice
numpy>=1.25.0,<3.0

